Move orguser files to admin-console/common
