+++ apps/desktop/tsconfig.json
@@ -10,7 +10,7 @@
     "types": [],
     "baseUrl": ".",
     "paths": {
-      "@bitwarden/admin-console": ["../../libs/admin-console/src"],
+      "@bitwarden/admin-console/common": ["../../libs/admin-console/src/common"],
       "@bitwarden/angular/*": ["../../libs/angular/src/*"],
       "@bitwarden/auth/common": ["../../libs/auth/src/common"],
       "@bitwarden/auth/angular": ["../../libs/auth/src/angular"],
