+++ libs/admin-console/src/common/organization-user/abstractions/index.ts
@@ -1,0 +1,1 @@
+export * from "./organization-user.service";
