+++ libs/admin-console/src/common/organization-user/models/requests/index.ts
@@ -6,3 +6,4 @@ export * from "./organization-user-invite.request";
 export * from "./organization-user-reset-password.request";
 export * from "./organization-user-reset-password-enrollment.request";
 export * from "./organization-user-update.request";
+export * from "./organization-user-bulk.request";
