+++ libs/admin-console/src/common/organization-user/models/requests/organization-user-accept-init.request.ts
@@ -1,4 +1,4 @@
-import { OrganizationKeysRequest } from "../../../models/request/organization-keys.request";
+import { OrganizationKeysRequest } from "@bitwarden/common/admin-console/models/request/organization-keys.request";
 
 export class OrganizationUserAcceptInitRequest {
   token: string;
