+++ libs/admin-console/src/common/organization-user/models/requests/organization-user-invite.request.ts
@@ -1,0 +1,12 @@
+import { OrganizationUserType } from "@bitwarden/common/admin-console/enums";
+import { PermissionsApi } from "@bitwarden/common/admin-console/models/api/permissions.api";
+import { SelectionReadOnlyRequest } from "@bitwarden/common/admin-console/models/request/selection-read-only.request";
+
+export class OrganizationUserInviteRequest {
+  emails: string[] = [];
+  type: OrganizationUserType;
+  accessSecretsManager: boolean;
+  collections: SelectionReadOnlyRequest[] = [];
+  groups: string[];
+  permissions: PermissionsApi;
+}
