+++ libs/admin-console/src/common/organization-user/models/requests/organization-user-reset-password-enrollment.request.ts
@@ -1,4 +1,4 @@
-import { SecretVerificationRequest } from "../../../../auth/models/request/secret-verification.request";
+import { SecretVerificationRequest } from "@bitwarden/common/auth/models/request/secret-verification.request";
 
 export class OrganizationUserResetPasswordEnrollmentRequest extends SecretVerificationRequest {
   resetPasswordKey: string;
