+++ libs/admin-console/src/common/organization-user/models/responses/organization-user-bulk.response.ts
@@ -1,4 +1,4 @@
-import { BaseResponse } from "../../../../models/response/base.response";
+import { BaseResponse } from "@bitwarden/common/models/response/base.response";
 
 export class OrganizationUserBulkResponse extends BaseResponse {
   id: string;
