+++ libs/admin-console/src/common/organization-user/services/index.ts
@@ -1,0 +1,1 @@
+export * from "./default-organization-user-api.service";
