--- libs/common/src/admin-console/abstractions/organization-user/requests/organization-user-invite.request.ts
@@ -1,12 +1,0 @@
-import { OrganizationUserType } from "../../../enums";
-import { PermissionsApi } from "../../../models/api/permissions.api";
-import { SelectionReadOnlyRequest } from "../../../models/request/selection-read-only.request";
-
-export class OrganizationUserInviteRequest {
-  emails: string[] = [];
-  type: OrganizationUserType;
-  accessSecretsManager: boolean;
-  collections: SelectionReadOnlyRequest[] = [];
-  groups: string[];
-  permissions: PermissionsApi;
-}
