--- libs/common/src/admin-console/abstractions/organization-user/requests/organization-user-update.request.ts
@@ -1,11 +1,0 @@
-import { OrganizationUserType } from "../../../enums";
-import { PermissionsApi } from "../../../models/api/permissions.api";
-import { SelectionReadOnlyRequest } from "../../../models/request/selection-read-only.request";
-
-export class OrganizationUserUpdateRequest {
-  type: OrganizationUserType;
-  accessSecretsManager: boolean;
-  collections: SelectionReadOnlyRequest[] = [];
-  groups: string[] = [];
-  permissions: PermissionsApi;
-}
