--- libs/common/src/admin-console/services/organization-user/requests/index.ts
@@ -1,1 +1,0 @@
-export * from "./organization-user-bulk.request";
